rename noise models file
