+++ src/smsfusion/noise/__init__.py
@@ -1,3 +1,3 @@
-from ._noise_models import gauss_markov, random_walk, white_noise
+from ._noise import gauss_markov, random_walk, white_noise
 
 __all__ = ["gauss_markov", "random_walk", "white_noise"]
